Implement enhanced garbage collection and memory management features
- Introduced `gc_free_object` to safely free allocated objects, improving memory handling.
- Updated `gc_sweep` to accept a parameter for full collection, allowing for more flexible garbage collection.
- Added `gc_collect_all` function to trigger complete garbage collection, enhancing memory cleanup capabilities.
- Modified the VM to utilize the new garbage collection function, ensuring efficient memory management during execution.
- Fixed object type formatting in `object_to_string` for better compatibility with different data types.

+++ src/object.c
@@ -63,7 +63,7 @@ DLLEXPORT char* object_to_string(object_t* _obj) {
         case OBJECT_TYPE_DOUBLE: {
             double intpart;
             if (modf(_obj->value.f64, &intpart) == 0.0) {
-                sprintf(str, "%d", (int) _obj->value.f64);
+                sprintf(str, "%ld", (long) _obj->value.f64);
             } else {
                 sprintf(str, "%.2lf", (double) _obj->value.f64);
             }
